refactor: 스키마 파일명 바뀐 컨벤션에 따른 변경 및 임포트문 수정 (#47)
