+++ src/hooks/queries/useAdminApplications.query.ts
@@ -1,10 +1,10 @@
 import {QUERY_KEYS} from '@/constants/query-keys';
-import {
+import type {
   GetAdminApplicationsParamsType,
   GetAdminApplicationsResponse,
-} from '@/schemas/admin/admin-applications-schema';
+} from '@/schemas/admin/admin-applications.schema';
 import {getAdminApplications} from '@/services/api/admin/admin.application.api';
-import {ErrorResponse} from '@/schemas/common/common-schema';
+import type {ErrorResponse} from '@/schemas/common/common-schema';
 import {useQuery} from '@tanstack/react-query';
 
 export const useAdminApplicationsQuery = (
