+++ src/hooks/queries/useAdminRecruitmentInformations.query.ts
@@ -1,7 +1,7 @@
 import {QUERY_KEYS} from '@/constants/query-keys';
-import {RecruitmentInformation} from '@/schemas/admin/admin-recruitment-information-schema';
+import type {RecruitmentInformation} from '@/schemas/admin/admin-recruitment-information.schema';
 import {getRecruitmentInformations} from '@/services/api/admin/admin.recruitment.info.api';
-import {ErrorResponse} from '@/schemas/common/common-schema';
+import type {ErrorResponse} from '@/schemas/common/common-schema';
 import {useQuery} from '@tanstack/react-query';
 
 export const useAdminRecruitmentInformationsQuery = (generationId: number) =>
